chore(playground): migrated demo app to hook plugin
